contrib/xot-implementation/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.3.6
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
polars==0.20.15
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
-e git+https://github.com/BrianLYS/taskgen.git@567764c849fbe71b7cbf6957d7aaaa6053f05bdb#egg=taskgen_ai
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
wcwidth==0.2.13
